reaction function and component, add comment function and component moved out from Confession
