+++ src/assets/data/data.jsx
@@ -156,6 +156,7 @@ const availablePollDurations = [
 
 const CONFESSION_CHAR_LIMIT = 280;
 const POLL_QUESTION_CHAR_LIMIT = CONFESSION_CHAR_LIMIT;
+const COMMENT_CHAR_LIMIT = CONFESSION_CHAR_LIMIT;
 const CHOICE_CHAR_LIMIT = 25;
 
 const howItWorks = [
@@ -203,4 +204,5 @@ export {
   CONFESSION_CHAR_LIMIT,
   POLL_QUESTION_CHAR_LIMIT,
   CHOICE_CHAR_LIMIT,
+  COMMENT_CHAR_LIMIT,
 };
