+++ src/components/Confession.jsx
@@ -10,139 +10,27 @@ import {
   Text,
   IconButton,
   Button,
-  InputGroup,
-  Input,
-  InputRightAddon,
   useDisclosure,
   Badge,
   Menu,
   MenuButton,
   MenuList,
   MenuItem,
-  InputLeftElement,
 } from "@chakra-ui/react";
-import {
-  HamburgerIcon,
-  ChatIcon,
-  DeleteIcon,
-  WarningTwoIcon,
-} from "@chakra-ui/icons";
+import { HamburgerIcon, DeleteIcon, WarningTwoIcon } from "@chakra-ui/icons";
 import color from "../styles/colors";
 import Comment from "./Comment";
 import { useState } from "react";
-import ReactionButton from "./ReactionButton";
+
 import ConfessionModal from "./ConfessionModal";
-import { availableReactions } from "../assets/data/data";
+import ReactionsToConfession from "./ReactionsToConfession";
 import moment from "moment";
-import {
-  doc,
-  updateDoc,
-  db,
-  arrayUnion,
-  increment,
-} from "../firebase/firebase";
 import useToastMessage from "../hooks/useToastMessage";
 import ConfessionText from "./ConfessionText";
 import Poll from "./Poll";
+import AddComment from "./AddComment";
 
-const COMMENT_CHAR_LIMIT = 280;
 const ALLOWED_VISIBLE_COMMENTS = 3;
-
-export const ReactionsToConfession = (props) => {
-  const { reactToConfession, reactions } = props;
-  return (
-    <Flex alignItems="center" justifyContent="space-around">
-      {availableReactions.map((item) => (
-        <ReactionButton
-          {...item}
-          key={item.id}
-          reactToConfession={reactToConfession}
-          reactionCount={reactions[item.title]}
-        />
-      ))}
-    </Flex>
-  );
-};
-
-export const AddComment = (props) => {
-  const {
-    addCommentToConfession,
-    commentIsDisabled,
-    setUserComment,
-    userComment,
-    isCommenting,
-    loggedInBatchYear,
-    setIsCommenting,
-    id,
-    resetComment,
-  } = props;
-
-  return (
-    <Box marginTop="10px">
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          addCommentToConfession(
-            setIsCommenting,
-            id,
-            userComment,
-            resetComment
-          );
-        }}
-      >
-        <InputGroup size="md" alignItems="center">
-          {userComment && (
-            <InputLeftElement pointerEvents="none" width="fit-content">
-              <Tag size="md" variant="subtle" colorScheme="purple">
-                {loggedInBatchYear}
-              </Tag>
-            </InputLeftElement>
-          )}
-
-          <Input
-            placeholder={
-              commentIsDisabled
-                ? "User has disabled the comments"
-                : "Add a comment..."
-            }
-            variant="flushed"
-            isDisabled={commentIsDisabled}
-            focusBorderColor={color.primary}
-            onChange={(event) => setUserComment(event.target.value)}
-            value={userComment}
-            maxLength={COMMENT_CHAR_LIMIT}
-            flex="1"
-            paddingLeft={userComment && "60px"}
-          />
-
-          <InputRightAddon backgroundColor="transparent" border="none">
-            {userComment && (
-              <Text fontSize="xs" as="span" color="blackAlpha.700">
-                {userComment.length}/{COMMENT_CHAR_LIMIT}
-              </Text>
-            )}
-            <IconButton
-              aria-label="Add Comment"
-              icon={<ChatIcon />}
-              colorScheme="purple"
-              variant="ghost"
-              isDisabled={!userComment}
-              isLoading={isCommenting}
-              onClick={() =>
-                addCommentToConfession(
-                  setIsCommenting,
-                  id,
-                  userComment,
-                  resetComment
-                )
-              }
-            />
-          </InputRightAddon>
-        </InputGroup>
-      </form>
-    </Box>
-  );
-};
 
 const Confession = (props) => {
   const [userComment, setUserComment] = useState("");
@@ -178,6 +66,7 @@ const Confession = (props) => {
     getConfessions,
     loggedInBatchYear,
     addCommentToConfession,
+    reactToConfession,
   } = props;
 
   const visibleComments = comments?.slice(0, ALLOWED_VISIBLE_COMMENTS);
@@ -198,64 +87,6 @@ const Confession = (props) => {
 
   const resetComment = () => {
     setUserComment("");
-  };
-
-  const userAlreadyReacted = () => {
-    let reactionArray;
-    reactionArray = localStorage.getItem("reactionList");
-    reactionArray = JSON.parse(reactionArray);
-
-    if (reactionArray && reactionArray.includes(id)) {
-      return true;
-    }
-
-    if (reactionArray) {
-      reactionArray.push(id);
-    } else {
-      reactionArray = [];
-      reactionArray.push(id);
-    }
-
-    localStorage.setItem("reactionList", JSON.stringify(reactionArray));
-
-    return false;
-  };
-
-  const reactToConfession = async (type) => {
-    const confessionRef = doc(db, "confessions", id);
-
-    if (userAlreadyReacted()) {
-      showToastMessage(
-        "Already Reacted",
-        "You have already reacted to this confession",
-        "warning"
-      );
-
-      return;
-    }
-
-    switch (type) {
-      case "like":
-        await updateDoc(confessionRef, {
-          "reactions.like": increment(1),
-        });
-
-        break;
-      case "funny":
-        await updateDoc(confessionRef, {
-          "reactions.funny": increment(1),
-        });
-
-        break;
-      case "shock":
-        await updateDoc(confessionRef, {
-          "reactions.shock": increment(1),
-        });
-
-        break;
-    }
-
-    getConfessions();
   };
 
   return (
@@ -336,6 +167,7 @@ const Confession = (props) => {
 
         <CardFooter flexDirection="column">
           <ReactionsToConfession
+            id={id}
             reactToConfession={reactToConfession}
             reactions={reactions}
           />
